Work around for an unknown issue with the org version straight
This reverts to the version bundled with the machine's emacs which on
both kronos and dyonisos happens to be the same one that gets picked
up by straight's mirror. I suppose that if I stick to the latest emacs
the version should be upgraded every now and then whihc is sufficient
for my needs.