Use maximized frames instead of fullscreen ones
This is because on macOS lsp-ui-doc creates invisible children frames
to show the documentation, which causes macOS to create another
fullscreen black frame and occasionaly move to it. Using a maximized
frame gets rid of this problem as the children frame are correctly
invisble. This does not matter on Linux setups as we use i3 over there anyways.