Clean up generating infix definers
There is still a need for a solution for the case where there are more
than one prefix key beyond the leader. Maybe the solution is to make
it a prefix definer proper by binding the respective keys for every
level in variables so that the prefix key string can be constructed
pragmatically.